Insert e Delete do  Consultorio prontos

+++ public/scripts/consultorio.js
@@ -80,26 +80,32 @@ const fecharModalButton = document.getElementById("fecharModal");
 fecharModalButton.addEventListener("click", fecharModalConclusao);
 
 // Captura o evento de clique do botão "Remover"
-document.addEventListener('click', function (event) {
-  if (event.target.classList.contains('btn-remove')) {
-    // Obtém o ID do consultório a ser removido
-    const consultorioId = event.target.dataset.id;
-
-    // Envia uma solicitação de exclusão para o servidor
-    fetch(`/consultorio/${consultorioId}`, {
-      method: 'DELETE',
-    })
-      .then(response => {
-        // Verifica se a solicitação foi bem-sucedida
-        if (response.ok) {
-          // Atualiza a página ou realiza qualquer outra ação necessária
-          window.location.reload();
-        } else {
-          console.error('Erro ao excluir o consultório');
-        }
-      })
-      .catch(error => {
-        console.error('Erro ao enviar a solicitação de exclusão',error);
-});
-}
-});
+document.addEventListener('click', function(event) {
+    if (event.target.classList.contains('btn-remove')) {
+      // Obtém o ID do consultório a ser removido
+      const consultorioId = event.target.dataset.id;
+      console.log(consultorioId)
+      // Verifica se o ID do consultório está definido
+      if (consultorioId) {
+        // Envia uma solicitação de exclusão para o servidor
+        fetch(`/consultorio/${consultorioId}`, {
+          method: 'DELETE',
+        })
+        .then(response => {
+          // Verifica se a solicitação foi bem-sucedida
+          if (response.ok) {
+            // Atualiza a página ou realiza qualquer outra ação necessária
+            window.location.reload();
+          } else {
+            console.error('Erro ao excluir o consultório');
+          }
+        })
+        .catch(error => {
+          console.error('Erro ao enviar a solicitação de exclusão', error);
+        });
+      } else {
+        console.error('ID do consultório não está definido');
+      }
+    }
+  });
+
